move meat of deploy complexity to separate Deploy module
